+++ bot/src/bot/bot.js
@@ -1,5 +1,9 @@
 // Основной файл бота Telegram. Использует dotenv, telegraf, сервисы задач,
 // загрузку файлов в R2 и JWT-аутентификацию.
+require('dotenv').config()
+if (process.env.NODE_ENV !== 'production') {
+  console.log('BOT_TOKEN:', process.env.BOT_TOKEN)
+}
 const { botToken, botApiUrl, appUrl, chatId, r2, port } = require('../config')
 const { Telegraf, Markup } = require('telegraf')
 const messages = require('../messages')